requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.15
django-cors-headers>=4.3
Pillow>=10.0
django-ratelimit>=4.1
